refac(portforward): code organization
