+++ crates/kftray-portforward/src/lib.rs
@@ -4,9 +4,14 @@ pub mod models;
 pub mod pod_finder;
 pub mod port_forward;
 
+mod tcp_forwarder;
+mod udp_forwarder;
+
 pub use core::*;
 
 pub use client::*;
 pub use models::*;
 pub use pod_finder::*;
 pub use port_forward::*;
+pub use tcp_forwarder::*;
+pub use udp_forwarder::*;
